hw7

+++ scrapper/src/main/java/io/swagger/api/LinksApiController.java
@@ -6,10 +6,14 @@ import edu.java.model.ApiException;
 import edu.java.model.LinkResponse;
 import edu.java.model.ListLinksResponse;
 import edu.java.model.RemoveLinkRequest;
+import io.github.bucket4j.Bandwidth;
+import io.github.bucket4j.Bucket;
+import io.github.bucket4j.Refill;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.net.URI;
+import java.time.Duration;
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import org.slf4j.Logger;
@@ -32,13 +36,20 @@ public class LinksApiController implements LinksApi {
     @Autowired
     private JdbcLinkService linkService;
     private final ObjectMapper objectMapper;
-
+    private final Bucket bucket;
     private final HttpServletRequest request;
+
+    private static final int ERROR = 404;
 
     @org.springframework.beans.factory.annotation.Autowired
     public LinksApiController(ObjectMapper objectMapper, HttpServletRequest request) {
         this.objectMapper = objectMapper;
         this.request = request;
+        Bandwidth limit =
+            Bandwidth.classic(2 * 2 * 2 * 2 + 2 * 2, Refill.greedy(2 * 2 * 2 * 2 + 2 * 2, Duration.ofMinutes(1)));
+        this.bucket = Bucket.builder()
+            .addLimit(limit)
+            .build();
     }
 
     @SuppressWarnings("MultipleStringLiterals")
@@ -48,37 +59,45 @@ public class LinksApiController implements LinksApi {
         @Parameter(in = ParameterIn.DEFAULT, description = "", required = true, schema = @Schema()) @Valid @RequestBody
         RemoveLinkRequest body
     ) {
-        try {
-            linkService.remove(tgChatId, body.getLink());
-        } catch (ApiException e) {
-            if (e.code == NOT_FOUND) {
-                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-            } else {
-                return new ResponseEntity(HttpStatus.CONFLICT);
+        if (bucket.tryConsume(1)) {
+            try {
+                linkService.remove(tgChatId, body.getLink());
+            } catch (ApiException e) {
+                if (e.getCode() == ERROR) {
+                    return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                } else {
+                    return new ResponseEntity(HttpStatus.CONFLICT);
+                }
             }
+            return new ResponseEntity(HttpStatus.OK);
+        } else {
+            return new ResponseEntity(HttpStatus.TOO_MANY_REQUESTS);
         }
-        return new ResponseEntity(HttpStatus.OK);
     }
 
     public ResponseEntity<ListLinksResponse> linksGet(
         @Parameter(in = ParameterIn.HEADER, description = "", required = true, schema = @Schema())
         @RequestHeader(value = "Tg-Chat-Id", required = true) Long tgChatId
     ) {
-        try {
-            var res = linkService.listAll(tgChatId);
-            ListLinksResponse response = new ListLinksResponse();
-            for (URI uri : res) {
-                var curr = new LinkResponse();
-                curr.setUrl(uri.toString());
-                response.addLinksItem(curr);
+        if (bucket.tryConsume(1)) {
+            try {
+                var res = linkService.listAll(tgChatId);
+                ListLinksResponse response = new ListLinksResponse();
+                for (URI uri : res) {
+                    var curr = new LinkResponse();
+                    curr.setUrl(uri.toString());
+                    response.addLinksItem(curr);
+                }
+                return new ResponseEntity<ListLinksResponse>(response, HttpStatus.OK);
+            } catch (ApiException e) {
+                if (e.getCode() == ERROR) {
+                    return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                } else {
+                    return new ResponseEntity(HttpStatus.CONFLICT);
+                }
             }
-            return new ResponseEntity<ListLinksResponse>(response, HttpStatus.OK);
-        } catch (ApiException e) {
-            if (e.code == NOT_FOUND) {
-                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-            } else {
-                return new ResponseEntity(HttpStatus.CONFLICT);
-            }
+        } else {
+            return new ResponseEntity(HttpStatus.TOO_MANY_REQUESTS);
         }
     }
 
@@ -88,16 +107,20 @@ public class LinksApiController implements LinksApi {
         @Parameter(in = ParameterIn.DEFAULT, description = "", required = true, schema = @Schema()) @Valid @RequestBody
         AddLinkRequest body
     ) {
-        try {
-            linkService.add(tgChatId, body.getLink());
-        } catch (ApiException e) {
-            if (e.code == NOT_FOUND) {
-                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-            } else {
-                return new ResponseEntity(HttpStatus.CONFLICT);
+        if (bucket.tryConsume(1)) {
+            try {
+                linkService.add(tgChatId, body.getLink());
+            } catch (ApiException e) {
+                if (e.getCode() == ERROR) {
+                    return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                } else {
+                    return new ResponseEntity(HttpStatus.CONFLICT);
+                }
             }
+            return new ResponseEntity(HttpStatus.OK);
+        } else {
+            return new ResponseEntity(HttpStatus.TOO_MANY_REQUESTS);
         }
-        return new ResponseEntity(HttpStatus.OK);
     }
 
 }
